feat: set up Express server with MongoDB connection and route handling
