+++ backend/src/db.ts
@@ -1,0 +1,12 @@
+import mongoose from 'mongoose';
+
+export const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.MONGODB_URI!);
+    console.log('Connected to MongoDB');
+  } catch (error) {
+    console.error('MongoDB connection error:', error);
+    process.exit(1);
+  }
+};
+
